# specs/001-mvp-mistake-notes/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 错题与笔记智能辅导 MVP API
  version: 1.0.0
  description: >
    API contract for the frontend MVP. All business endpoints require
    Bearer JWT except /auth/wx-login.
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /notes/ingest:
    post:
      tags: [Notes]
      summary: OCR and classify uploaded note file, returning candidates only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteIngestRequest'
      responses:
        "200":
          description: Note candidate for review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteIngestResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "502":
          $ref: '#/components/responses/ExternalServiceError'
  /notes:
    post:
      tags: [Notes]
      summary: Persist a user-reviewed note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteCreateRequest'
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDetail'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Notes]
      summary: List notes for current user
      parameters:
        - $ref: '#/components/parameters/SubjectId'
        - $ref: '#/components/parameters/TagId'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, pending, trashed]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        "200":
          description: Paginated note list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteListResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /notes/{id}:
    get:
      tags: [Notes]
      summary: Get note detail owned by current user
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        "200":
          description: Note detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDetail'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Notes]
      summary: Update reviewed note fields and tags
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpdateRequest'
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDetail'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Notes]
      summary: Soft delete a note
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        "204":
          description: Soft deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SubjectId:
      name: subject_id
      in: query
      schema:
        type: integer
        minimum: 1
...
    TagId:
      name: tag_id
      in: query
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unsupportedType:
              value:
                error:
                  code: UPLOAD_UNSUPPORTED_TYPE
                  message: 文件格式不支持
                  request_id: req_123
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            authRequired:
              value:
                error:
                  code: AUTH_REQUIRED
                  message: 请先登录
                  request_id: req_123
    NotFound:
      description: Resource not found or not owned by current user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ExternalServiceError:
      description: External OCR/LLM/WeChat dependency failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorCode:
      type: string
      enum:
        - AUTH_REQUIRED
        - TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - OBJECT_NOT_FOUND
        - VALIDATION_ERROR
        - UPLOAD_UNSUPPORTED_TYPE
        - UPLOAD_LIMIT_EXCEEDED
        - PDF_PARSE_FAILED
        - PDF_TEXT_EXTRACTION_FAILED
        - UNSUPPORTED_MIME_TYPE
        - OCR_FAILED
        - LLM_SCHEMA_INVALID
        - RATE_LIMIT_EXCEEDED
        - CONFLICT
        - EXTERNAL_SERVICE_ERROR
        - RESTORE_WINDOW_EXPIRED
        - INTERNAL_ERROR
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              example: 请求无效
            request_id:
              type: string
              example: req_123
            details:
              type: object
              additionalProperties: true
...
    SourceType:
      type: string
      enum: [photo, pdf]
    ItemStatus:
      type: string
      enum: [active, pending]
...
    IngestStatus:
      type: string
      enum: [ready, pending_classification, ocr_failed]
    TagKind:
      type: string
      enum: [knowledge_point, question_type, error_cause]
...
    TagInput:
      type: object
      required: [kind, name]
      properties:
        kind:
          $ref: '#/components/schemas/TagKind'
        name:
          type: string
          example: 二次函数顶点式
    Tag:
      allOf:
        - $ref: '#/components/schemas/TagInput'
        - type: object
          required: [id, normalized_name]
          properties:
            id:
              type: string
              format: uuid
            normalized_name:
              type: string
              example: 顶点式
...
    NoteCandidate:
      type: object
      required: [subject_id, content, knowledge_points]
      properties:
        subject_id:
          type: integer
        content:
          type: string
        knowledge_points:
          type: array
          items:
            type: string
        page_number:
          type: integer
          nullable: true
    NoteIngestRequest:
      type: object
      required: [object_key, source_type, subject_id]
      properties:
        object_key:
          type: string
        source_type:
          $ref: '#/components/schemas/SourceType'
        subject_id:
          type: integer
    NoteIngestResponse:
      type: object
      required: [session_id, status, ocr_text, candidates]
      properties:
        session_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/IngestStatus'
        ocr_text:
          type: string
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/NoteCandidate'
        error_code:
          $ref: '#/components/schemas/ErrorCode'
          nullable: true
    NoteCreateRequest:
      type: object
      required: [subject_id, source_type, object_key, content, tags]
      properties:
        ingest_session_id:
          type: string
          format: uuid
          nullable: true
        subject_id:
          type: integer
        source_type:
          $ref: '#/components/schemas/SourceType'
        object_key:
          type: string
        preview_object_key:
          type: string
          nullable: true
        ocr_text:
          type: string
          default: ""
        content:
          type: string
        status:
          $ref: '#/components/schemas/ItemStatus'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagInput'
    NoteUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/NoteCreateRequest'
      description: Partial update in implementation; schema documents allowed fields.
    NoteListItem:
      type: object
      required: [id, subject_id, content_excerpt, tags, created_at]
      properties:
        id:
          type: string
          format: uuid
        subject_id:
          type: integer
        thumbnail_url:
          type: string
          format: uri
          nullable: true
        content_excerpt:
          type: string
        status:
          $ref: '#/components/schemas/ItemStatus'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    NoteDetail:
      allOf:
        - $ref: '#/components/schemas/NoteListItem'
        - type: object
          required: [object_key, ocr_text, content]
          properties:
            object_key:
              type: string
            file_url:
              type: string
              format: uri
              nullable: true
            ocr_text:
              type: string
            content:
              type: string
    NoteListResponse:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NoteListItem'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer